[동일] 역경매 상세보기 뷰 수정 등...
1. 역경매 상세보기 모달 수정 (지도 나옴)
- rest 추가
2. 역경매 본인 삭제만 가능하도록 함
3. 역경매 날자 검색시 '>' 에서 '>=' 로 변경

+++ src/main/java/com/stay/alive/auction/reverse/service/ReverseauctionService.java
@@ -1,8 +1,6 @@
 package com.stay.alive.auction.reverse.service;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -62,10 +60,11 @@ public class ReverseauctionService {
 		return reverseauctionMapper.updateReverseauction(reverseauction);
 	}
 	// 7. 역경매 삭제
-	public int removeReverseauction(int reverseauctionNo) {
+	public int removeReverseauction(int reverseauctionNo, String memberId) {
 		System.out.println("ReverseauctionService.removeReverseauction()");
 		int i=0;
-		if(!reverseauctionMapper.selectReverseauctionState(reverseauctionNo).equals("낙찰완료")) {
+		if(!reverseauctionMapper.selectReverseauctionState(reverseauctionNo).equals("낙찰완료")
+				&& reverseauctionMapper.selectReverseauctionOne(reverseauctionNo).getMemberId().equals(memberId)) {
 			i+=reverseauctionTenderMapper.deleteReverseauctionTender(reverseauctionNo);
 			i+=reverseauctionMapper.deleteReverseauction(reverseauctionNo);
 		}
